Move 'shlex_process_stream' into utils
This eliminates 'parsing.process_stdin', which now holds a single
function and is no longer even devoted to parsing stdin. As
'shlex_process_stream' is a relatively small and simple method, move
it into utils.
