+++ src/globus_cli/utils.py
@@ -1,5 +1,6 @@
 import inspect
 import json
+import shlex
 from typing import Callable, Dict, Iterable, Iterator, List, Optional
 
 import click
@@ -135,3 +136,28 @@ class PagingWrapper:
             self._step()
             yield cur
             yielded += 1
+
+
+def shlex_process_stream(process_command, stream):
+    """
+    Use shlex to process stdin line-by-line.
+    Also prints help text.
+
+    Requires that @process_command be a Click command object, used for
+    processing single lines of input. helptext is prepended to the standard
+    message printed to interactive sessions.
+    """
+    # use readlines() rather than implicit file read line looping to force
+    # python to properly capture EOF (otherwise, EOF acts as a flush and
+    # things get weird)
+    for line in stream.readlines():
+        # get the argument vector:
+        # do a shlex split to handle quoted paths with spaces in them
+        # also lets us have comments with #
+        argv = shlex.split(line, comments=True)
+        if argv:
+            try:
+                process_command.main(args=argv)
+            except SystemExit as e:
+                if e.code != 0:
+                    raise
